Add support for LRB traces to the simulator
LRB offers a 17gb cdn trace (2.8B events) that it was analyzed against.

The archive support was improved so that a tar.gz can be supplied and
the entries will be concatenated.

When hit/miss penalty support was added, the Optimal policy degraded
by having to hold the entire event in-memory. This is now conditioned
so that for non-latency traces only the key (primitive long) is held.
This reduces GC pressure and improves performance on large traces.